handle pulling data of multiple tables
